+++ src/module/scale.rs
@@ -7,12 +7,18 @@ use crate::module::{DrawerModule, Module, Slider};
 use crate::text::Svg;
 
 pub struct Scale {
+    default_scale: f64,
     scale: f64,
 }
 
 impl Scale {
-    pub fn new() -> Self {
-        Self { scale: 2. }
+    pub fn new() -> Option<Self> {
+        let msg = IpcMessage::Scale { scale: None, app_id: None };
+        let scale = match catacomb_ipc::send_message(&msg) {
+            Ok(Some(IpcMessage::ScaleReply { scale: WindowScale::Fixed(scale) })) => scale,
+            _ => return None,
+        };
+        Some(Self { scale, default_scale: scale })
     }
 }
 
@@ -24,8 +30,8 @@ impl Module for Scale {
 
 impl Slider for Scale {
     fn set_value(&mut self, value: f64) -> Result<()> {
-        // Map from `0..=1` to `1..=3`.
-        let mut scale = value * 2. + 1.;
+        // Limit scale to within one above/below the default.
+        let mut scale = value * 2. + self.default_scale - 1.;
 
         // Round to nearest multiple of .5.
         scale = (scale * 2.).round() / 2.;
@@ -38,14 +44,14 @@ impl Slider for Scale {
 
     fn on_touch_up(&mut self) -> Result<()> {
         // Update Catacomb's scale.
-        let msg = IpcMessage::Scale { scale: WindowScale::Fixed(self.scale), app_id: None };
+        let msg = IpcMessage::Scale { scale: Some(WindowScale::Fixed(self.scale)), app_id: None };
         catacomb_ipc::send_message(&msg)?;
         Ok(())
     }
 
     fn value(&self) -> f64 {
-        // Map back from `1..=3` to `0..=1`.
-        (self.scale - 1.) / 2.
+        // Map back to within one above/below the default.
+        (self.scale - self.default_scale + 1.) / 2.
     }
 
     fn svg(&self) -> Svg {